move quaternion definitions to core
rename coordinateutils/quaternion.h coordinateutils/pointing.h
