+++ coordinateutils/CMakeLists.txt
@@ -13,14 +13,13 @@ add_library(coordinateutils SHARED
   src/FlatSkyMap.cxx
   src/mapdata.cxx
   src/maputils.cxx
-  src/quaternion.cxx
+  src/pointing.cxx
   src/python.cxx
 )
 
 target_link_libraries(coordinateutils core)
 execute_process(COMMAND ln -fsn ${CMAKE_CURRENT_SOURCE_DIR}/python ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/coordinateutils)
 
-add_spt3g_test(quattest)
 add_spt3g_test(quatangtest)
 add_spt3g_test(transtest SLOWTEST)
 add_spt3g_test(skymappybindings)
